+++ packages/protractor/spec/screenplay/interactions/UseAngular.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from '@integration/testing-tools';
-import { Ensure, equals } from '@serenity-js/assertions';
+import { Ensure, equals, isTrue } from '@serenity-js/assertions';
 import { actorCalled, Question } from '@serenity-js/core';
 
 import { protractor } from 'protractor';
@@ -32,7 +32,7 @@ describe('UseAngular', function () {
             /** @test {UseAngular} */
             it('can be enabled', () => actorCalled('Bernie').attemptsTo(
                 UseAngular.enableSynchronisation(),
-                Ensure.that(IsSynchronisationEnabled(), equals(true)),
+                Ensure.that(IsSynchronisationEnabled(), isTrue()),
             ));
         });